round 2 almost works
» Shortened wordArray and filled in some items for round2 to test the continue button
» Created an on-click for the continue button, that will hide the modal box and add the round2 items to wordArray

RESULT: This mostly works. I learned that I had to use apply() with push() to prevent the second array from nesting in the first. I want the second array's elements to be pushed into the first, not the single element of the array itself. However, after several test runs, the last word from the first round keeps getting repeated as the first word of the second round, and the last word in the second array, der Zucker, never appears. I think this has to do with the pop() method I used earlier?