categories .slideDown()
Most likely going to use .slideDown() to reveal the categories section. I wanted to avoid manually setting the #categories-wrapper top property equal to the header's height, since I'm planning on shrinking the height for mobile devices.

I originally tried creating a variable and setting it equal to $('header').height(), which was problematic for a couple of reasons. One, it returns only a number, so a unit (e.g., px) would have to be concatenated. Two, it returns only the height of the header's content, so the padding height was excluded.

Setting the variable equal to $('header').css('height'), instead, resolved both problems presented by jQuery's .height(). Ultimately, the slide down method is more fun—and it doesn't require any height/top position values.