category selection update
When I made the checkboxes visible, to confirm that the selection rings corresponded with check status, I noticed that there were inconsistencies. I got rid of the if statement conditionals and the css style updates, and just used toggle to make the selection ring visible when the boxes are checked. So far, this seems to work consistently.