categories not yet working
Not yet sure how to get the categories to work. I'm able to group all the words from the first two rounds, but I need to then clear wordArray and fill it with words from the categories that are checked. Seems really roundabout to use concat. Will keep trying tomorrow.