#open-categories: word total
Had the idea to push the array lengths into an arrLengths array, yesterday, but needed to search MDN to remind myself of the reduce() method. Am now able to generate the total number of words in the gameArray without having to manually update it each time I add new rounds.