miniReviewArray created
Set up the miniReviewArray (and the endReviewArray), and updated the #answer-key onclick, so any time a user uses the answer key, the current word will be added to the miniReviewArray. I also added a forEach statement to the #mini-review onclick, so now each word in the miniReviewArray will appear in the end-of-round review box.