buildMiniReview()
If the hint or answer buttons were pressed more than once, or if both the hint and answer button were pressed to solve a word, the solution was duplicated in the mini array. I'm using jQuery's $.uniqueSort() function to remove duplicates, so I created buildMiniReview() to push a randomWord into miniReviewArray and then uniqueSort the array. This function is called in the #hint-bulb and #answer-key click events.