trying to shrink the input selection
I started to add more checked-input if statements, but I also want to figure out a way to reduce this. So far, I tried replacing the individual IDs with "this," while also replacing the category names with $(this).attr('name'). I thought it might be easier to have the IDs match the category names, to see if I can avoid writing out almost the same code for each input. It works, but for only one category at a time.