red quiz buttons
Updated the .quiz-buttons click event to turn quiz buttons red if an incorrect answer is selected. The quiz-bzzt keyframes animation is applied by adding the class .add-quiz-bzzt, and it's followed by the addition of .red-button after a small delay. (The latter class ensures that the "wrong" buttons reject hover styles, so they'll stay red while the other buttons turn yellow on hover.)