checkForArticles() is named
I turned the block of code that filters der, die, das quiz words into the named function checkForArticles(), and I reused it in the .lightning-round onclick. I also had to update the .open-categories onclick, to filter out the words without articles before the word total is tallied in the quiz page's lightning round.