accessibility update
In an effort to make the site more accessible, I added tabindex="0" to .open-categories and the lightning-round and select-categories buttons, so they'd be keyboard focusable. I also added tabindex="1" to the three quiz buttons and to the next item on the nav panel, since there was an issue making the elements in the .open-categories section focusable.