input.category[type=checkbox] on click update
Moved some things around to make a general "if this is checked" statement, rather than maintaining a statement for each of the input buttons.

The words now get pushed into the new customRoundArray when the appropriate category is selected, and if any category is selected, wordArray is cleared. If the user clicks the Select Categories button, the customRoundArray words get pushed into wordArray.