consolidatedArray update
Since consolidatedArray is built by concatenating gameArray, it was being mutated by "gameArray.splice(1)" in the .select-categories on click. This is what prevented users from adding new categories to the Customized Round after an initial selection had been made. My solution was to move consolidatedArray out of the #open-categories on click and into the global scope.

I also updated the .select-categories on click to check whether selectionConfirmed is true or false. If false, wordArray is emptied before the customRoundArray words are pushed in. If true, wordArray isn't zeroed out, as subsequent confirmations would remove the customRoundArray words.